requirements.txt
bokeh==3.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
Flask==2.2.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kafka==1.3.5
kafka-python==2.0.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
py4j==0.10.9.5
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
tornado==6.2
tzdata==2023.3
urllib3==1.26.15
websocket==0.2.1
Werkzeug==2.2.3
xyzservices==2023.2.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
